# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CampusConnect REST API Specification",
    "description": "Comprehensive OpenAPI 3.0.3 contract for National University of Computer & Emerging Sciences (FAST / NUCES) Campus Management System.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Express API Server"
    },
    {
      "url": "https://api.campusconnect.nu.edu.pk/api",
      "description": "Production API Server"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Register Student Account",
        "description": "Enforces @nu.edu.pk institutional domain validation.",
        "responses": {
          "201": { "description": "Student account created successfully" },
          "400": { "description": "Invalid payload or non-FAST email domain" },
          "409": { "description": "Email address already registered" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate User Session",
        "description": "Issues HttpOnly JWT cookie on successful credential verification.",
        "responses": {
          "200": { "description": "User authenticated successfully" },
          "401": { "description": "Invalid credentials or account deactivated" }
        }
      }
    },
...
    "/auth/logout-all": {
      "post": {
        "summary": "Revoke All Multi-Device User Sessions",
        "description": "Increments session_version in PostgreSQL, invalidating active JWTs across all devices.",
        "responses": {
          "200": { "description": "All active multi-device sessions revoked" },
          "401": { "description": "Unauthenticated or session revoked" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get Authenticated User Profile",
        "responses": {
          "200": { "description": "Authenticated user object" },
          "401": { "description": "Unauthenticated or session revoked" }
        }
      }
    },
...
    "/messages/conversations": {
      "post": {
        "summary": "Initiate Marketplace Listing Conversation",
        "responses": {
          "201": { "description": "Conversation created or retrieved" },
          "400": { "description": "Cannot message own listing or invalid payload" }
        }
      },
      "get": {
        "summary": "List Student Marketplace Conversations",
        "responses": { "200": { "description": "Active student conversations list" } }
      }
    },
    "/messages/conversations/{id}/messages": {
      "get": {
        "summary": "Fetch Conversation Message History",
        "responses": {
          "200": { "description": "Paginated conversation messages" },
          "403": { "description": "Forbidden - Non-participant request rejected" }
        }
      },
      "post": {
        "summary": "Send Message via REST Fallback",
        "responses": {
          "201": { "description": "Message persisted and dispatched" },
          "403": { "description": "Forbidden - Non-participant request rejected" }
        }
      }
    },
...
  }
}